Use definite library versions
In platformio.ini, the symbol ^ has been removed for all referenced libraries (parameter lib_deps). This is a workaround to avoid problems with changed interfaces in newer library versions.

For the ...schreibfaul1/ESP32-audioI2S library, the version #d9a2421 has been added.

A better (but more time-consuming) solution would be to adapt the code to current library versions.